Moved source_reader.py in hazardlib
