+++ openquake/commonlib/__init__.py
@@ -18,4 +18,4 @@
 
 # keep the following line to have a __version__
 from openquake.baselib import __version__
-from openquake.hazardlib import logictree
+from openquake.hazardlib import logictree, source_reader
