+++ openquake/hazardlib/source_reader.py
